refactor: update `authorId` field type to `Long` across community entities and services
- Changed `authorId` from `Integer` to `Long` in `Article`, `Recruitment`, and relevant DTOs, services, and repositories.
- Updated database schema accordingly to reflect type change.
- Adjusted method signatures and type usage in controllers and services for consistency.

+++ src/main/java/kr/ac/koreatech/sw/kosp/domain/community/article/model/Article.java
@@ -87,7 +87,7 @@ public class Article extends BaseEntity {
         return board.getId();
     }
 
-    public Integer getAuthorId() {
+    public Long getAuthorId() {
         return author.getId();
     }
